refactored icon folder name to components
