+++ app/components/Question/Components/EditIcon.js
