+++ app/components/Question/Components/StarIcon.js
