+++ app/components/Question/Components/ThinkIcon.js
