+++ app/components/Question/Question.js
@@ -2,9 +2,9 @@
  * Created by will on 10/03/18.
  */
 import React from 'react'
-import StarIcon from './icons/StarIcon';
-import ThinkIcon from './icons/ThinkIcon';
-import EditIcon from './icons/EditIcon';
+import StarIcon from './Components/StarIcon';
+import ThinkIcon from './Components/ThinkIcon';
+import EditIcon from './Components/EditIcon';
 import styles from './styles.css';
 
 export default () => {
